+++ Smudgecraft-Heroes-Skills/SkillSlash/src/me/marwzoor/skillslash/SkillSlash.java
@@ -131,7 +131,7 @@ public class SkillSlash extends ActiveSkill
 												
 						double damage = event.getDamage()*sEffect.getPercent();
 						
-						event.setDamage((int) damage);
+						event.setDamage(damage);
 						
 						LivingEntity le = (LivingEntity) event.getEntity();
